// packages/analytics-js-plugins/src/utilities/retryQueue/types.d.ts
import { DoneCallback } from '../../types/plugins';
export type QueueBackoff = {
  minRetryDelay: number;
  maxRetryDelay: number;
  factor: number;
  jitter: number;
};
export type QueueTimeouts = {
  ackTimer: number;
  reclaimTimer: number;
  reclaimTimeout: number;
  reclaimWait: number;
};
export type InProgressQueueItem = {
  item: Record<string, any> | string | number | Record<string, any>[] | string[] | number[];
  done: DoneCallback;
  attemptNumber: number;
};